- Implemented DTOs and mapping layer to decouple domain from API
- Refactored controller and service methods to return ResponseEntity
- Standardized method names for clarity and consistency
- Adjusted database schema and renamed table/columns accordingly
- Removed ID from exposed DTO to improve API design

+++ src/main/java/dev/samir/backangulart/Crud/Controller/ClothController.java
@@ -1,10 +1,10 @@
 package dev.samir.backangulart.Crud.Controller;
 
-import dev.samir.backangulart.Crud.Model.Cloth;
+import dev.samir.backangulart.Crud.DtoMapper.ClothDto;
 import dev.samir.backangulart.Crud.Service.ClothService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
 
 @RestController
@@ -18,30 +18,55 @@ public class ClothController {
     }
 
     @GetMapping("/show")
-    public List<Cloth> showList() {
-        return clothService.showlist();
+    public ResponseEntity<List<ClothDto>>  showList() {
+        List<ClothDto> cloths = clothService.showlist();
+        return ResponseEntity.ok().body(cloths) ;
     }
 
     @GetMapping("/show/{id}")
-    public ResponseEntity<Cloth> showListById(@PathVariable Long id) {
-        return clothService.showListById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<?> showListById(@PathVariable Long id) {
+        ClothDto clothById = clothService.showListById(id);
+        if (clothById != null) {
+            return ResponseEntity.ok(clothById);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Cloth with id " + id + " not found");
+        }
     }
 
     @PostMapping("/add")
-    public Cloth create(@RequestBody Cloth cloth) {
-        return clothService.create(cloth);
+    public ResponseEntity<String> create(@RequestBody ClothDto clothDto) {
+        ClothDto createCloth = clothService.create(clothDto);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Cloth created successfully: " + createCloth.name());
     }
 
     @PutMapping("/update/{id}")
-    public Cloth update(@PathVariable Long id, @RequestBody Cloth cloth) {
-        return clothService.update(id, cloth);
+    public ResponseEntity<?> update(@PathVariable Long id, @RequestBody ClothDto clothDto) {
+
+        ClothDto updateCloth = clothService.update(id, clothDto);
+
+        if (updateCloth != null) {
+            String msg = "Cloth updated successfully: "
+                    + updateCloth.name() +
+                    " | Size: " + updateCloth.size() +
+                    " | Color: " + updateCloth.color();
+            return ResponseEntity.ok(msg);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Cloth with id " + id + " not found");
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id) {
-        clothService.delete(id);
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        if (clothService.showListById(id) != null) {
+            clothService.delete(id);
+            return ResponseEntity.ok("Cloth deleted successfully!" + id);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Cloth with id " + id + " not found");
+        }
     }
 
 }
